fixed file tree in macOS
