+++ clients/apple/SwiftLockbookCore/Sources/SwiftLockbookCore/Models/File.swift
@@ -1,5 +1,4 @@
 import Foundation
-import SwiftUI
 
 public struct File: Codable, Identifiable, Equatable, Hashable, Comparable {
     public var fileType: FileType
