+++ clients/apple/macOS/DesktopHomeView.swift
@@ -5,10 +5,12 @@ import SwiftWorkspace
 
 struct DesktopHomeView: View {
     @State var searchInput: String = ""
+    @State var expandedFolders: [File] = []
+    @State var lastOpenDoc: File? = nil
 
     var body: some View {
         NavigationView {
-            SidebarView(searchInput: $searchInput)
+            SidebarView(searchInput: $searchInput, expandedFolders: $expandedFolders, lastOpenDoc: $lastOpenDoc)
                 .searchable(text: $searchInput, prompt: "Search")
             
             DetailView()
@@ -18,11 +20,12 @@ struct DesktopHomeView: View {
 
 struct SidebarView: View {
     @EnvironmentObject var search: SearchService
+    
     @Environment(\.isSearching) var isSearching
 
     @Binding var searchInput: String
-    @State var expandedFolders: [File] = []
-    @State var lastOpenDoc: File? = nil
+    @Binding var expandedFolders: [File]
+    @Binding var lastOpenDoc: File?
     
     @State var treeBranchState: Bool = true
