+++ clients/apple/macOS/TreeView/FileTreeView.swift
@@ -22,6 +22,7 @@ struct FileTreeView: NSViewRepresentable {
         if treeView.numberOfColumns != 1 {
             delegate.documentSelected = { meta in
                 if meta.fileType == .Document {
+                    print("opening \(meta.name)")
                     DI.workspace.requestOpenDoc(meta.id)
                 } else if meta.fileType == .Folder {
                     DI.workspace.selectedFolder = meta.id
